+++ src/VariableStatementAnalysis.ts
@@ -1,9 +1,10 @@
 import { CodeLocation } from "./CodeLocation";
+import { VariableTypes, VariableValues } from "./mapStack";
 
 export interface VariableStatementAnalysis {
     name: string;
-    value: number | boolean;
-    variableType: string;
+    value: VariableValues;
+    variableType: VariableTypes;
     text: string;
     expressionLocation: CodeLocation;
     identifierLocation: CodeLocation;